fix ScalaJS
