--- core-tests/shared/src/test/scala/java/time/OffsetDateTime.scala
@@ -1,5 +1,0 @@
-package java.time
-
-trait OffsetDateTime
-
-object OffsetDateTime
