--- core-tests/shared/src/test/scala/java/time/ZoneId.scala
@@ -1,5 +1,0 @@
-package java.time
-
-trait ZoneId
-
-object ZoneId
