--- core-tests/shared/src/test/scala/java/time/ZoneOffset.scala
@@ -1,5 +1,0 @@
-package java.time
-
-trait ZoneOffset
-
-object ZoneOffset
